src/BarcodeReaderFunctionsProject/requirements.txt
pyzbar==0.1.8
azure-functions
Pillow
requests